Refactor Join screen into few reusable components
