+++ src/component/Header.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Header = () => (
+  <h1>Copyplash</h1>
+)
+
+export { Header }
