+++ src/component/Loading.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Loading = () => (
+  <h3>Loading...</h3>
+)
+
+export { Loading }
