+++ src/component/index.js
@@ -1,0 +1,2 @@
+export * from './Header'
+export * from './Loading'
